requirements.txt
APScheduler==3.7.0
PyYAML==5.4.1
speedtest-cli==2.1.3
